lin and win generation

+++ power.sh
@@ -1,3 +1,9 @@
+NAME=$1
+if [ "$#" == "0" ]; then
+    echo "No arguments provided"
+    exit 1
+fi
+
 JRE_VERSION="1.7.0_40"
 OUT="rap"
 rm -rf $OUT
@@ -8,9 +14,11 @@ for f in `ls drivers`; do
     mkdir $OUT/$f
     mkdir $OUT/$f/core
     mkdir $OUT/$f/core/JRE
+    
     cp -r JRE/$JRE_VERSION/$f/* $OUT/$f/core/JRE
+    WD=`pwd`
     cd drivers/$f
-    ./build.sh
+    bash build.sh $1 "$WD/$OUT/$f/core"
     cd ../..
   fi
 done
